+++ clickable/utils.py
@@ -7,6 +7,7 @@ import glob
 import inspect
 from os.path import dirname, basename, isfile, isdir, join
 
+from jsonschema import validate, ValidationError
 import yaml
 
 from clickable.builders.base import Builder
@@ -14,15 +15,6 @@ from clickable.logger import logger
 from clickable.exceptions import FileNotFoundException, ClickableException
 from clickable.config.constants import Constants
 from clickable.logger import Colors
-
-SCHEMA_VALIDATOR_AVAILABLE = True
-try:
-    from jsonschema import validate, ValidationError
-except ImportError:
-    SCHEMA_VALIDATOR_AVAILABLE = False
-
-
-# TODO use these subprocess functions everywhere
 
 
 def prepare_command(cmd, shell=False):
@@ -90,7 +82,7 @@ def find_pattern(pattern, base_dir, exclude_dir=None):
     return files
 
 
-def find(
+def find(  # pylint: disable=too-many-positional-arguments
     names,
     cwd,
     temp_dir=None,
@@ -205,6 +197,9 @@ def env(name):
 
 
 def get_builder(config, container, debug_build=False):
+    if not config.builder:
+        raise ClickableException(
+            "Builder requested when there is none defined. Please report this bug.")
     builder_classes = get_builders()
     return builder_classes[config.builder](config, container, debug_build)
 
@@ -270,32 +265,33 @@ def load_config_schema(name):
 
 
 def validate_config_format(config, schema, name, path):
-    if SCHEMA_VALIDATOR_AVAILABLE:
-        try:
-            validate(instance=config, schema=schema)
-        except ValidationError as e:
-            logger.error('The %s config file "%s" contains invalid fields!', name, path)
-            error_message = e.message
-            # Lets add the key to the invalid value
-            if e.path:
-                if len(e.path) > 1 and isinstance(e.path[-1], int):
-                    error_message = f'{error_message} (in "{e.path[-2]}")'
-                else:
-                    error_message = f'{error_message} (in "{e.path[-1]}")'
-            raise ClickableException(error_message) from e
-    else:
-        logger.warning('Dependency "jsonschema" not found. Could not validate config file.')
+    try:
+        validate(instance=config, schema=schema)
+    except ValidationError as e:
+        logger.error('The %s config file "%s" contains invalid fields!', name, path)
+        error_message = e.message
+        # Lets add the key to the invalid value
+        if e.path:
+            if len(e.path) > 1 and isinstance(e.path[-1], int):
+                error_message = f'{error_message} (in "{e.path[-2]}")'
+            else:
+                error_message = f'{error_message} (in "{e.path[-1]}")'
+        raise ClickableException(error_message) from e
 
 
-def pull_image(image, skip_existing=True):
-    if not skip_existing or not image_exists(image):
-        docker_executable = get_docker_command()
+def pull_image(image, skip_existing=True, docker_executable=None):
+    if not skip_existing or not image_exists(image, docker_executable):
+        if not docker_executable:
+            docker_executable = get_docker_command()
+
         command = f'{docker_executable} pull {image}'
         run_subprocess_check_call(command)
 
 
-def image_exists(image):
-    docker_executable = get_docker_command()
+def image_exists(image, docker_executable=None):
+    if not docker_executable:
+        docker_executable = get_docker_command()
+
     command = f'{docker_executable} image inspect {image}'
     return run_subprocess_call(
         command,
@@ -304,16 +300,30 @@ def image_exists(image):
     ) == 0
 
 
-def image_based_on(image, base):
-    docker_executable = get_docker_command()
-    command_template = f'{docker_executable} history --no-trunc -q'
-    command_base = f'{command_template} {base}'
-    command_image = f'{command_template} {image}'
+def get_image_hash(image, docker_executable=None):
+    if not docker_executable:
+        docker_executable = get_docker_command()
 
-    hash_base = run_subprocess_check_output(command_base).split('\n', 1)[0]
-    history = run_subprocess_check_output(command_image).strip()
+    format_string = "'{{.Id}}'"
+    command = f"{docker_executable} inspect --format {format_string} {image}"
 
-    return hash_base in history
+    return run_subprocess_check_output(command).strip()
+
+
+def image_based_on_hash(image, base_hash, docker_executable=None):
+    if not docker_executable:
+        docker_executable = get_docker_command()
+
+    format_string = "'{{.Config.Labels.base_image_hash}}'"
+    command = f"{docker_executable} inspect --format {format_string} {image}"
+
+    hash_label = run_subprocess_check_output(command).strip()
+
+    return base_hash == hash_label
+
+
+def image_based_on(image, base, docker_executable=None):
+    return image_based_on_hash(image, get_image_hash(base, docker_executable), docker_executable)
 
 
 def makedirs(path):